Update dependencies and enhance AES-GCM and BIKE implementations
- Upgraded the OQS dependency from version 0.8.0 to 0.11.0 for improved cryptographic features.
- Added hex encoding/decoding support to the project.
- Refactored AES-CTR implementation to utilize type aliases for better clarity.
- Introduced streaming decryption functionality for AES-GCM, including initialization, update, and finalization methods.
- Enhanced the BIKE decapsulation process with detailed comments and improved error handling.
- Updated tests and examples to reflect changes in the AES and BIKE modules, ensuring robustness and clarity.

+++ src/dilithium/compression.rs
@@ -302,7 +302,7 @@ fn medium_compression(signature: &[u8], variant: DilithiumVariant) -> Result<Vec
         } 
         // Check if delta is small (can be encoded in 4 bits)
         else if (byte as i16 - prev_byte as i16).abs() < 8 {
-            let delta = (byte as i8 - prev_byte as i8) & 0x0F;
+            let delta = ((byte as i16 - prev_byte as i16) as i8) & 0x0F;
             
             // If we have a previous delta waiting, combine them
             if compressed.last() == Some(&0xFE) {
